@@ -27,58 +27,62 @@
   ],
   "scripts": {
     "compile": "npx hardhat clean && npx hardhat compile",
-    "deploy": "bash ./scripts/deploy.sh",
     "docs": "scripts/prepare-docs.sh",
-    "fullcheck": "yarn lint && yarn tsc && yarn tslint && yarn slither",
+    "fullCheck": "yarn lint && yarn tsc && yarn eslint && yarn cspell && yarn slither",
     "hooks": "git config core.hooksPath .githooks || true",
+    "no-hooks": "git config core.hooksPath .git/hooks",
     "lint": "npx solhint \"contracts/**/*.sol\"",
     "prepare": "yarn compile && yarn hooks",
     "slither": "slither .",
     "test": "yarn tsc && npx hardhat test",
     "tsc": "tsc --noEmit",
-    "tslint": "npx tslint --project ."
+    "eslint": "npx eslint --cache --ext .js,.jsx,.ts,.tsx .",
+    "cspell": "npx cspell \"**/*\""
   },
   "dependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.0.2",
-    "@nomiclabs/hardhat-web3": "^2.0.0",
-    "@openzeppelin/contracts": "^3.4.1",
-    "@openzeppelin/contracts-ethereum-package": "^3.0.0",
-    "@openzeppelin/hardhat-upgrades": "^1.9.0",
-    "@skalenetwork/skale-manager-interfaces": "^0.1.2",
-    "axios": "^0.21.1",
-    "dotenv": "^10.0.0",
-    "ethereumjs-util": "^7.0.10",
-    "ethers": "^5.3.1",
+    "@nomiclabs/hardhat-ethers": "^2.0.4",
+    "@openzeppelin/contracts": "^4.4.2",
+    "@openzeppelin/contracts-upgradeable": "^4.4.2",
+    "@openzeppelin/hardhat-upgrades": "^1.13.0",
+    "@skalenetwork/skale-manager-interfaces": "^1.0.0-develop.0",
+    "@typechain/hardhat": "^4.0.0",
+    "axios": "^0.25.0",
+    "dotenv": "^14.3.2",
+    "ethereumjs-util": "^7.1.3",
+    "ethers": "^5.5.3",
     "ethers-eip712": "^0.2.0",
-    "hardhat": "^2.4.1",
-    "web3": "^1.3.6"
+    "hardhat": "^2.8.3"
   },
   "devDependencies": {
-    "@nomiclabs/hardhat-etherscan": "^2.1.1",
-    "@typechain/ethers-v5": "^5.0.0",
-    "@types/chai": "^4.2.14",
+    "@nomiclabs/hardhat-etherscan": "^3.0.0",
+    "@nomiclabs/hardhat-waffle": "^2.0.2",
+    "@typechain/ethers-v5": "^9.0.0",
+    "@types/chai": "^4.3.0",
     "@types/chai-almost": "^1.0.1",
     "@types/chai-as-promised": "^7.1.3",
+    "@types/elliptic": "^6.4.14",
     "@types/minimist": "^1.2.0",
-    "@types/mocha": "^8.2.0",
-    "@types/node": "^14.14.25",
-    "@types/sinon-chai": "^3.2.5",
-    "bignumber.js": "^9.0.0",
-    "chai": "^4.2.0",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.12",
+    "@types/sinon-chai": "^3.2.8",
+    "@typescript-eslint/eslint-plugin": "^5.10.1",
+    "@typescript-eslint/parser": "^5.10.1",
+    "bignumber.js": "^9.0.2",
+    "chai": "^4.3.6",
     "chai-almost": "^1.0.1",
     "chai-as-promised": "^7.1.1",
+    "cspell": "^5.17.0",
     "elliptic": "^6.5.3",
+    "eslint": "^8.7.0",
     "ethereum-waffle": "^3.2.2",
     "ganache-cli": "^6.11.0",
-    "hardhat-typechain": "^0.3.4",
-    "solhint": "^3.2.1",
-    "solidity-coverage": "^0.7.14",
-    "solidity-docgen": "^0.5.7",
+    "kill-port": "^1.6.1",
+    "solhint": "3.3.6",
+    "solidity-coverage": "^0.7.18",
+    "solidity-docgen": "^0.5.16",
     "ts-generator": "^0.1.1",
-    "ts-node": "9.0.0",
-    "tslint": "^6.1.3",
-    "tslint-no-unused-expression-chai": "^0.1.4",
-    "typechain": "^4.0.1",
-    "typescript": "4.0.3"
+    "ts-node": "10.4.0",
+    "typechain": "^7.0.0",
+    "typescript": "4.5.5"
   }
 }
