@@ -1,19 +1,18 @@
 #!/usr/bin/env node
 
-const path = require('path');
-const proc = require('child_process');
-const startCase = require('lodash.startcase');
+import { parse } from 'path';
+import { execFileSync } from 'child_process';
 
 const baseDir = process.argv[2];
 const PREFIX = "/technology/skale-manager/"
 
-const files = proc.execFileSync(
+const files = execFileSync(
   'find', [baseDir, '-type', 'f'], { encoding: 'utf8' }
 ).split('\n').filter(s => s !== '');
 
 const links = files.map((file) => {
   const doc = file.replace(baseDir + "/", '');
-  const { dir, name } = path.parse(doc);
+  const { dir, name } = parse(doc);
   return {
     title: name,
     link: `${PREFIX}${dir ? dir + '/' : ''}${name}`.toLowerCase(),
