@@ -1,41 +1,25 @@
 import { deployContractManager } from "../test/tools/deploy/contractManager";
 import { deployValidatorService } from "../test/tools/deploy/delegation/validatorService";
 import { deploySkaleManager } from "../test/tools/deploy/skaleManager";
-import { ContractManager, Schains, SkaleManager, ValidatorService } from "../typechain";
+import { deploySchainsInternal } from "../test/tools/deploy/schainsInternal";
+import { ContractManager, Schains, SkaleManager, ValidatorService } from "../typechain-types";
 import { privateKeys } from "../test/tools/private-keys";
-import * as elliptic from "elliptic";
 import { deploySchains } from "../test/tools/deploy/schains";
 import fs from 'fs';
-import { ethers, web3 } from "hardhat";
+import { ethers } from "hardhat";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
-import { BigNumberish, Event } from "ethers";
+import { Event, Wallet } from "ethers";
+import { getPublicKey, getValidatorIdSignature } from "../test/tools/signatures";
+import { fastBeforeEach } from "../test/tools/mocha";
+import { SchainType } from "../test/tools/types";
+import { TypedEvent } from "../typechain-types/common";
+import { SchainNodesEvent } from "../typechain-types/ISchains";
 
-const ec = new elliptic.ec("secp256k1");
-
-async function getValidatorIdSignature(validatorId: BigNumberish, signer: SignerWithAddress) {
-    const hash = web3.utils.soliditySha3(validatorId.toString());
-    if (hash) {
-        let signature = await web3.eth.sign(hash, signer.address);
-        signature = (
-            signature.slice(130) === "00" ?
-            signature.slice(0, 130) + "1b" :
-            (
-                signature.slice(130) === "01" ?
-                signature.slice(0, 130) + "1c" :
-                signature
-            )
-        );
-        return signature;
-    } else {
-        return "";
-    }
-}
-
-function findEvent(events: Event[] | undefined, eventName: string) {
+function findEvent<TargetEvent extends TypedEvent>(events: Event[] | undefined, eventName: string) {
     if (events) {
         const target = events.find((event) => event.event === eventName);
         if (target) {
-            return target;
+            return target as TargetEvent;
         } else {
             throw new Error("Event was not emitted");
         }
@@ -47,49 +31,51 @@ function findEvent(events: Event[] | undefined, eventName: string) {
 describe("createSchains", () => {
     let owner: SignerWithAddress;
     let validator: SignerWithAddress;
-    let node: SignerWithAddress;
+    let nodeAddress: Wallet
 
     let contractManager: ContractManager;
     let validatorService: ValidatorService;
     let skaleManager: SkaleManager;
     let schains: Schains;
 
-    beforeEach(async () => {
-        [owner, validator, node] = await ethers.getSigners();
+    fastBeforeEach(async () => {
+        [owner, validator] = await ethers.getSigners();
+        nodeAddress = new Wallet(String(privateKeys[2])).connect(ethers.provider);
+        await owner.sendTransaction({to: nodeAddress.address, value: ethers.utils.parseEther("10000")});
         contractManager = await deployContractManager();
 
         validatorService = await deployValidatorService(contractManager);
         skaleManager = await deploySkaleManager(contractManager);
         schains = await deploySchains(contractManager);
+        await deploySchainsInternal(contractManager);
     });
 
     it("gas based on nodes amount", async () => {
-        const validatorId = 1;
-
+        await validatorService.grantRole(await validatorService.VALIDATOR_MANAGER_ROLE(), owner.address);
         await validatorService.connect(validator).registerValidator("Validator", "", 0, 0);
+        const validatorId = await validatorService.getValidatorId(validator.address);
         await validatorService.disableWhitelist();
-        const signature = await getValidatorIdSignature(validatorId, node);
-        await validatorService.connect(validator).linkNodeAddress(node.address, signature);
+        const signature = await getValidatorIdSignature(validatorId, nodeAddress);
+        await validatorService.connect(validator).linkNodeAddress(nodeAddress.address, signature);
         await schains.grantRole(await schains.SCHAIN_CREATOR_ROLE(), owner.address);
 
         const maxNodesAmount = 1000;
         const gasLimit = 12e6;
         const measurements = []
-        const publicKey = ec.keyFromPrivate(String(privateKeys[2]).slice(2)).getPublic();
         for (let nodeId = 0; nodeId < maxNodesAmount; ++nodeId) {
-            await skaleManager.connect(node).createNode(
+            await skaleManager.connect(nodeAddress).createNode(
                 1, // port
                 0, // nonce
                 "0x7f" + ("000000" + nodeId.toString(16)).slice(-6), // ip
                 "0x7f" + ("000000" + nodeId.toString(16)).slice(-6), // public ip
-                ["0x" + publicKey.x.toString('hex'), "0x" + publicKey.y.toString('hex')], // public key
-                "d2-" + nodeId, // name)
-                "somedomain.name");
+                getPublicKey(nodeAddress),
+                `d2-${nodeId}`, // name)
+                "some.domain.name");
 
             const nodesAmount = nodeId + 1;
             if (nodesAmount >= 16) {
-                const result = await (await schains.addSchainByFoundation(0, 1, 0, "schain-" + nodeId, owner.address)).wait();
-                const nodeInGroup = findEvent(result.events, "SchainNodes").args?.nodesInGroup;
+                const result = await (await schains.addSchainByFoundation(0, SchainType.SMALL, 0, `schain-${nodeId}`, owner.address, ethers.constants.AddressZero, [])).wait();
+                const nodeInGroup = findEvent<SchainNodesEvent>(result.events, "SchainNodes").args?.nodesInGroup;
                 console.log("Nodes in Schain:");
                 const setOfNodes = new Set();
                 for (const nodeOfSchain of nodeInGroup) {
