@@ -2,6 +2,10 @@ import axios from "axios";
 import { TypedDataUtils } from "ethers-eip712";
 import * as ethUtil from 'ethereumjs-util';
 import chalk from "chalk";
+import type { ethers } from "ethers";
+import { HardhatEthersHelpers } from "@nomiclabs/hardhat-ethers/types";
+
+type Ethers = typeof ethers & HardhatEthersHelpers;
 
 enum Network {
     MAINNET = 1,
@@ -30,6 +34,56 @@ const URLS = {
     }
 }
 
+interface SafeInfoResponse{
+    address: string,
+    nonce: number,
+    threshold: number,
+    owners:	string[],
+    masterCopy:	string,
+    modules: string[],
+    fallbackHandler: string,
+    guard: string,
+    version: string
+}
+
+interface SafeMultisigEstimateTx{
+    safe: string,
+    to:	string,
+    value: number,
+    data?: string
+    operation: number
+    gasToken?: string
+}
+
+interface SafeMultisigEstimateTxResponseV2 {
+    safeTxGas: string,
+    baseGas: string,
+    dataGas: string,
+    operationalGas: string,
+    gasPrice: string,
+    lastUsedNonce: number,
+    gasToken: string,
+    refundReceiver: string
+}
+
+interface SafeMultisigTransaction {
+    safe: string,
+    to: string,
+    value: number,
+    data?: string,
+    operation: number,
+    gasToken?: string,
+    safeTxGas: number,
+    baseGas: number,
+    gasPrice: number,
+    refundReceiver?: string,
+    nonce: number,
+    contractTransactionHash: string,
+    sender:	string,
+    signature?:	string,
+    origin?: string
+}
+
 function getMultiSendAddress(chainId: number) {
     if (chainId === Network.MAINNET) {
         return ADDRESSES.multiSend[chainId];
@@ -38,7 +92,7 @@ function getMultiSendAddress(chainId: number) {
     } else if ([Network.GANACHE, Network.HARDHAT].includes(chainId)) {
         return ZERO_ADDRESS;
     } else {
-        throw Error("Can't get multiSend contract at network with chainId = " + chainId);
+        throw Error(`Can't get multiSend contract at network with chainId = ${chainId}`);
     }
 }
 
@@ -48,7 +102,7 @@ export function getSafeTransactionUrl(chainId: number) {
     } else if (chainId === 4) {
         return URLS.safe_transaction[chainId];
     } else {
-        throw Error("Can't get safe-transaction url at network with chainId = " + chainId);
+        throw Error(`Can't get safe-transaction url at network with chainId = ${chainId}`);
     }
 }
 
@@ -58,7 +112,7 @@ export function getSafeRelayUrl(chainId: number) {
     } else if (chainId === 4) {
         return URLS.safe_relay[chainId];
     } else {
-        throw Error("Can't get safe-relay url at network with chainId = " + chainId);
+        throw Error(`Can't get safe-relay url at network with chainId = ${chainId}`);
     }
 }
 
@@ -72,23 +126,24 @@ function concatTransactions(transactions: string[]) {
     }).join("");
 }
 
-export async function createMultiSendTransaction(ethers: any, safeAddress: string, privateKey: string, transactions: string[]) {
-    const chainId: number = (await ethers.provider.getNetwork()).chainId;
+export async function createMultiSendTransaction(ethers: Ethers, safeAddress: string, privateKey: string, transactions: string[]) {
+    const chainId = (await ethers.provider.getNetwork()).chainId;
     const multiSendAddress = getMultiSendAddress(chainId);
     const multiSendAbi = [{"constant":false,"inputs":[{"internalType":"bytes","name":"transactions","type":"bytes"}],"name":"multiSend","outputs":[],"payable":false,"stateMutability":"nonpayable","type":"function"}];
     const multiSend = new ethers.Contract(multiSendAddress, new ethers.utils.Interface(multiSendAbi), ethers.provider);
 
     let nonce = 0;
     try {
-        const nonceResponse = await axios.get(`${getSafeTransactionUrl(chainId)}/api/v1/safes/${safeAddress}/`);
+        const nonceResponse = await axios.get<SafeInfoResponse>(`${getSafeTransactionUrl(chainId)}/api/v1/safes/${safeAddress}/`);
         nonce = nonceResponse.data.nonce;
     } catch (e) {
-        if (!e.toString().startsWith("Error: Can't get safe-transaction url")) {
+        if (!(e instanceof Error) || !e.toString().startsWith("Error: Can't get safe-transaction url")) {
             throw e;
         }
     }
 
     const tx = {
+        "safe": safeAddress,
         "to": multiSend.address,
         "value": 0, // Value in wei
         "data": multiSend.interface.encodeFunctionData("multiSend", [ concatTransactions(transactions) ]),
@@ -136,7 +191,7 @@ export async function createMultiSendTransaction(ethers: any, safeAddress: strin
     const { r, s, v } = ethUtil.ecsign(ethUtil.toBuffer(digestHex), privateKeyBuffer);
     const signature = ethUtil.toRpcSig(v, r, s).toString();
 
-    const txToSend = {
+    const txToSend: SafeMultisigTransaction = {
         ...tx,
         "contractTransactionHash": digestHex,  // Contract transaction hash calculated from all the field
         // Owner of the Safe proposing the transaction. Must match one of the signatures
@@ -148,19 +203,13 @@ export async function createMultiSendTransaction(ethers: any, safeAddress: strin
     return txToSend;
 }
 
-export async function sendSafeTransaction(safe: string, chainId: number, safeTx: any) {
+export async function sendSafeTransaction(safe: string, chainId: number, safeTx: SafeMultisigTransaction) {
     try {
         console.log("Estimate gas");
-        const estimateRequest = (({
-            to,
-            value,
-            data,
-            operation,
-            gasToken
-        }) => ({ to, value, data, operation, gasToken }))(safeTx);
+        const estimateRequest: SafeMultisigEstimateTx = safeTx;
 
         try {
-            const estimateResponse = await axios.post(
+            const estimateResponse = await axios.post<SafeMultisigEstimateTxResponseV2>(
                 `${getSafeRelayUrl(chainId)}/api/v2/safes/${safe}/transactions/estimate/`,
                 estimateRequest
             );
@@ -168,17 +217,19 @@ export async function sendSafeTransaction(safe: string, chainId: number, safeTx:
                 parseInt(estimateResponse.data.safeTxGas, 10) + parseInt(estimateResponse.data.baseGas, 10)}`));
         } catch (e) {
             console.log(chalk.red("Failed to estimate gas"));
-            console.log(e.toString());
+            console.log(e);
         }
 
         console.log(chalk.green("Send transaction to gnosis safe"));
-        await axios.post(`${getSafeTransactionUrl(chainId)}/api/v1/safes/${safe}/transactions/`, safeTx)
+        await axios.post(`${getSafeTransactionUrl(chainId)}/api/v1/safes/${safe}/multisig-transactions/`, safeTx)
     } catch (e) {
-        if (e.response) {
-            console.log(JSON.stringify(e.response.data, null, 4))
-            console.log(chalk.red(`Request failed with ${e.response.status} code`));
-        } else {
-            console.log(chalk.red("Request failed with unknown reason"));
+        if (axios.isAxiosError(e)) {
+            if (e.response) {
+                console.log(JSON.stringify(e.response.data, null, 4))
+                console.log(chalk.red(`Request failed with ${e.response.status} code`));
+            } else {
+                console.log(chalk.red("Request failed with unknown reason"));
+            }
         }
         throw e;
     }
