@@ -3,16 +3,14 @@ import {
     ConstantsHolder,
     BountyV2,
     NodesMock,
-    SkaleManager,
     DelegationController,
     SkaleToken,
     ValidatorService
-} from "../typechain";
-
+} from "../typechain-types";
 import { deployContractManager } from "./tools/deploy/contractManager";
 import { deployConstantsHolder } from "./tools/deploy/constantsHolder";
 import { deployBounty } from "./tools/deploy/bounty";
-import { currentTime, skipTime, skipTimeToDate } from "./tools/time";
+import { currentTime, nextMonth, skipTime, skipTimeToDate } from "./tools/time";
 import chaiAsPromised from "chai-as-promised";
 import chaiAlmost from "chai-almost";
 import * as chai from "chai";
@@ -21,44 +19,15 @@ import { deploySkaleToken } from "./tools/deploy/skaleToken";
 import { deployDelegationController } from "./tools/deploy/delegation/delegationController";
 import { deployValidatorService } from "./tools/deploy/delegation/validatorService";
 import { deployDelegationPeriodManager } from "./tools/deploy/delegation/delegationPeriodManager";
-import { deployDistributor } from "./tools/deploy/delegation/distributor";
 import { deploySkaleManagerMock } from "./tools/deploy/test/skaleManagerMock";
-import { privateKeys } from "./tools/private-keys";
-import * as elliptic from "elliptic";
-import { ethers, web3 } from "hardhat";
-import { solidity } from "ethereum-waffle";
+import { ethers } from "hardhat";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
-import { BigNumber, Event, ContractFactory } from "ethers";
+import { BigNumber } from "ethers";
 import { deployPunisher } from "./tools/deploy/delegation/punisher";
-
-import { deployLibraries, getLinkedContractFactory } from "../test/tools/deploy/factory";
+import { fastBeforeEach } from "./tools/mocha";
 
 chai.should();
 chai.use(chaiAsPromised);
-chai.use(solidity);
-const EC = elliptic.ec;
-const ec = new EC("secp256k1");
-
-function hexValue(value: string) {
-    if (value.length % 2 === 0) {
-        return value;
-    } else {
-        return "0" + value;
-    }
-}
-
-function findEvent(events: Event[] | undefined, eventName: string) {
-    if (events) {
-        const target = events.find((event) => event.event === eventName);
-        if (target) {
-            return target;
-        } else {
-            throw new Error("Event was not emitted");
-        }
-    } else {
-        throw new Error("Event was not emitted");
-    }
-}
 
 function getBountyForEpoch(epoch: number) {
     const bountyForFirst6Years = [385000000, 346500000, 308000000, 269500000, 231000000, 192500000];
@@ -68,18 +37,6 @@ function getBountyForEpoch(epoch: number) {
     } else {
         return bountyForFirst6Years[5] / 2 ** (Math.floor((year - 6) / 3) + 1);
     }
-}
-
-async function getContractFactoryWithLibraries(e: any, contractName: string) {
-    const libraryNames = [];
-    for (const str of e.toString().split(".sol:")) {
-        const libraryName = str.split("\n")[0];
-        libraryNames.push(libraryName);
-    }
-    libraryNames.shift();
-    const libraries = await deployLibraries(libraryNames);
-    const contractFactory = await getLinkedContractFactory(contractName, libraries);
-    return contractFactory;
 }
 
 describe("Bounty", () => {
@@ -96,14 +53,10 @@ describe("Bounty", () => {
 
     const ten18 = BigNumber.from(10).pow(18);
     const day = 60 * 60 * 24;
-    const month = 31 * day;
 
-    before(async() => {
+    fastBeforeEach(async () => {
         chai.use(chaiAlmost(2));
         [owner, admin, hacker, validator, validator2] = await ethers.getSigners();
-    });
-
-    beforeEach(async () => {
         contractManager = await deployContractManager();
         constantsHolder = await deployConstantsHolder(contractManager);
         bountyContract = await deployBounty(contractManager);
@@ -133,14 +86,14 @@ describe("Bounty", () => {
         await bountyContract.disableBountyReduction();
     });
 
-    describe("when validator is registered and has active delegations", async () => {
+    describe("when validator is registered and has active delegations", () => {
         let skaleToken: SkaleToken;
         let delegationController: DelegationController;
         let validatorService: ValidatorService;
 
         const validatorId = 1;
         const validatorAmount = 1e6;
-        beforeEach(async () => {
+        fastBeforeEach(async () => {
             skaleToken = await deploySkaleToken(contractManager);
             delegationController = await deployDelegationController(contractManager);
             validatorService = await deployValidatorService(contractManager);
@@ -152,7 +105,7 @@ describe("Bounty", () => {
             await validatorService.enableValidator(validatorId);
             await delegationController.connect(validator).delegate(validatorId, ten18.mul(validatorAmount).toString(), 2, "");
             await delegationController.connect(validator).acceptPendingDelegation(0);
-            await skipTime(ethers, month);
+            await nextMonth(contractManager);
         });
 
         async function calculateBounty(nodeId: number) {
@@ -164,10 +117,10 @@ describe("Bounty", () => {
             return bounty;
         }
 
-        describe("when second validator is registered and has active delegations", async () => {
+        describe("when second validator is registered and has active delegations", () => {
             const validator2Id = 2;
             const validator2Amount = 0.5e6;
-            beforeEach(async () => {
+            fastBeforeEach(async () => {
                 const delegationPeriodManager = await deployDelegationPeriodManager(contractManager);
                 const DELEGATION_PERIOD_SETTER_ROLE = await delegationPeriodManager.DELEGATION_PERIOD_SETTER_ROLE();
                 await delegationPeriodManager.grantRole(DELEGATION_PERIOD_SETTER_ROLE, owner.address);
@@ -178,10 +131,10 @@ describe("Bounty", () => {
                 await delegationPeriodManager.setDelegationPeriod(12, 200);
                 await delegationController.connect(validator2).delegate(validator2Id, ten18.mul(validator2Amount).toString(), 12, "");
                 await delegationController.connect(validator2).acceptPendingDelegation(1);
-                await skipTime(ethers, month);
+                await nextMonth(contractManager);
 
-                await skipTimeToDate(ethers, 1, 0); // Jan 1st
-                await constantsHolder.setLaunchTimestamp(await currentTime(web3));
+                await skipTimeToDate(1, 0); // Jan 1st
+                await constantsHolder.setLaunchTimestamp(await currentTime());
                 await constantsHolder.setMSR(ten18.mul(validator2Amount).toString());
             });
 
@@ -189,7 +142,7 @@ describe("Bounty", () => {
                 await nodes.registerNodes(2, validatorId);
                 await nodes.registerNodes(1, validator2Id);
 
-                await skipTime(ethers, 29 * day);
+                await skipTime(29 * day);
                 const bounty0 = await calculateBounty(0) + await calculateBounty(1);
                 const bounty1 = await calculateBounty(2);
                 bounty0.should.be.equal(bounty1);
@@ -204,7 +157,7 @@ describe("Bounty", () => {
                 const million = ten18.mul(1e6).toString();
 
                 // Jan 1st
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // validator1:
                 //     delegations:
@@ -218,10 +171,10 @@ describe("Bounty", () => {
 
                 await delegationController.connect(validator).requestUndelegation(0);
 
-                await skipTimeToDate(ethers, 15, 0);
+                await skipTimeToDate(15, 0);
 
                 // Jan 15th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -236,10 +189,10 @@ describe("Bounty", () => {
                 await delegationController.connect(validator2).delegate(2, million, 2, "");
                 await delegationController.connect(validator2).acceptPendingDelegation(2);
 
-                await skipTimeToDate(ethers, 30, 0);
+                await skipTimeToDate(30, 0);
 
                 // Jan 30th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -254,10 +207,10 @@ describe("Bounty", () => {
 
                 await punisher.slash(1, million);
 
-                await skipTimeToDate(ethers, 1, 1);
+                await skipTimeToDate(1, 1);
 
                 // Feb 1st
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -275,10 +228,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(0)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 15, 1);
+                await skipTimeToDate(15, 1);
 
                 // Feb 15th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -298,10 +251,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(0)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 27, 1);
+                await skipTimeToDate(27, 1);
 
                 // Feb 27th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -323,10 +276,10 @@ describe("Bounty", () => {
                 let bounty = await calculateBounty(0);
                 bounty.should.be.almost(getBountyForEpoch(0) + getBountyForEpoch(1));
 
-                await skipTimeToDate(ethers, 1, 2);
+                await skipTimeToDate(1, 2);
 
                 // March 1st
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -347,10 +300,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(0)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 15, 2);
+                await skipTimeToDate(15, 2);
 
                 // March 15th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -371,10 +324,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(0)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 29, 2);
+                await skipTimeToDate(29, 2);
 
                 // March 29th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -402,10 +355,10 @@ describe("Bounty", () => {
                 bounty.should.be.almost(bountyLeft * effectiveDelegated2 / (effectiveDelegated1 + effectiveDelegated2));
                 totalBounty += bounty;
 
-                await skipTimeToDate(ethers, 1, 3);
+                await skipTimeToDate(1, 3);
 
                 // April 1st
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -429,10 +382,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(1)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 15, 3);
+                await skipTimeToDate(15, 3);
 
                 // April 15th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -459,10 +412,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(2)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 28, 3);
+                await skipTimeToDate(28, 3);
 
                 // April 28th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -497,10 +450,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(2)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 1, 4);
+                await skipTimeToDate(1, 4);
 
                 // May 1st
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -531,10 +484,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(3)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 15, 4);
+                await skipTimeToDate(15, 4);
 
                 // May 15th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -580,10 +533,10 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(3)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 29, 4);
+                await skipTimeToDate(29, 4);
 
                 // May 29th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -631,10 +584,10 @@ describe("Bounty", () => {
 
                 totalBounty.should.be.lessThan(getBountyForEpoch(4));
 
-                await skipTimeToDate(ethers, 16, 5);
+                await skipTimeToDate(16, 5);
 
                 // June 16th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -667,14 +620,14 @@ describe("Bounty", () => {
                 await bountyContract.calculateBounty(3)
                     .should.be.eventually.rejectedWith("Transaction is sent too early");
 
-                await skipTimeToDate(ethers, 27, 5);
+                await skipTimeToDate(27, 5);
 
                 await delegationController.connect(validator2).requestUndelegation(6);
 
-                await skipTimeToDate(ethers, 28, 5);
+                await skipTimeToDate(28, 5);
 
                 // June 28th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -703,10 +656,10 @@ describe("Bounty", () => {
                 bounty.should.be.almost(0); // stake is too low
                 totalBounty += bounty;
 
-                await skipTimeToDate(ethers, 29, 6);
+                await skipTimeToDate(29, 6);
 
                 // July 29th
-                // console.log("ts: current", new Date(await currentTime(web3) * 1000));
+                // console.log("ts: current", new Date(await currentTime() * 1000));
 
                 // 1. validator1:
                 //     delegations:
@@ -741,62 +694,5 @@ describe("Bounty", () => {
                 totalBounty += bounty;
             });
         });
-
-        // this test was used to manually check bounty distribution
-
-        // it("30 nodes by 1 each day", async () => {
-        //     const nodesCount = 30;
-        //     const result = new Map<number, object[]>();
-        //     const queue = []
-        //     for (let i = 0; i < nodesCount; ++i) {
-        //         await nodes.registerNodes(1, validatorId);
-        //         console.log("Node", i, "is registered", new Date(await currentTime(web3) * 1000))
-        //         await skipTime(ethers, day);
-        //         result.set(i, []);
-        //         queue.push({nodeId: i, getBountyTimestamp: (await bountyContract.getNextRewardTimestamp(i)).toNumber()})
-        //     }
-        //     let minBounty = Infinity;
-        //     let maxBounty = 0;
-        //     const startTime = await currentTime(web3);
-        //     queue.sort((a, b) => {
-        //         return b.getBountyTimestamp - a.getBountyTimestamp;
-        //     });
-        //     for (let timestamp = startTime; timestamp < startTime + 365 * day; timestamp = await currentTime(web3)) {
-        //         const nodeInfo: {nodeId: number, getBountyTimestamp: number} | undefined = queue.pop();
-        //         assert(nodeInfo !== undefined);
-        //         if (nodeInfo) {
-        //             const nodeId = nodeInfo.nodeId;
-        //             if (timestamp < nodeInfo.getBountyTimestamp) {
-        //                 await skipTime(ethers, nodeInfo.getBountyTimestamp - timestamp);
-        //                 timestamp = await currentTime(web3)
-        //             }
-        //             console.log("Node", nodeId, new Date(await currentTime(web3) * 1000))
-        //             const bounty = web3.utils.toBN((await bountyContract.calculateBounty.call(nodeId))).div(ten18).toNumber();
-        //             // total[nodeIndex] += bounty;
-        //             await bountyContract.calculateBounty(nodeId);
-        //             await nodes.changeNodeLastRewardDate(nodeId);
-
-        //             nodeInfo.getBountyTimestamp = (await bountyContract.getNextRewardTimestamp(nodeId)).toNumber();
-        //             queue.push(nodeInfo)
-        //             queue.sort((a, b) => {
-        //                 return b.getBountyTimestamp - a.getBountyTimestamp;
-        //             });
-
-        //             minBounty = Math.min(minBounty, bounty);
-        //             maxBounty = Math.max(maxBounty, bounty);
-        //             result.get(nodeId)?.push({timestamp, bounty});
-        //         } else {
-        //             assert(false, "Internal error");
-        //         }
-        //     }
-        //     console.log(minBounty, maxBounty);
-        //     console.log(JSON.stringify(Array.from(result)));
-        //     const epochs = []
-        //     const timeHelpers = await deployTimeHelpers(contractManager);
-        //     for (let i = 0; i < 30; ++i) {
-        //         epochs.push((await timeHelpers.monthToTimestamp(i)).toNumber())
-        //     }
-        //     console.log(JSON.stringify(Array.from(epochs)));
-        // })
     });
 });
