@@ -1,16 +1,15 @@
 import { ConstantsHolder,
-         ContractManager } from "../typechain";
+         ContractManager } from "../typechain-types";
 import { deployConstantsHolder } from "./tools/deploy/constantsHolder";
 import { deployContractManager } from "./tools/deploy/contractManager";
 import { ethers } from "hardhat";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
 import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
-import { solidity } from "ethereum-waffle";
+import { fastBeforeEach } from "./tools/mocha";
 
 chai.should();
 chai.use(chaiAsPromised);
-chai.use(solidity);
 
 describe("ContractManager", () => {
   let user: SignerWithAddress;
@@ -18,7 +17,7 @@ describe("ContractManager", () => {
   let contractManager: ContractManager;
   let constantsHolder: ConstantsHolder;
 
-  beforeEach(async () => {
+  fastBeforeEach(async () => {
     [, user] = await ethers.getSigners();
 
     contractManager = await deployContractManager();
@@ -26,7 +25,7 @@ describe("ContractManager", () => {
   });
 
   it("Should add a right contract address (ConstantsHolder) to the register", async () => {
-    const simpleContractName: string = "Constants";
+    const simpleContractName = "Constants";
     await contractManager.connect(user).setContractsAddress(simpleContractName, constantsHolder.address)
       .should.be.eventually.rejectedWith("Ownable: caller is not the owner");
     await contractManager.setContractsAddress(simpleContractName, constantsHolder.address);
