@@ -1,27 +1,21 @@
 import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
-import { Decryption } from "../typechain";
-
-import { gasMultiplier } from "./tools/command_line";
-import { skipTime } from "./tools/time";
-
-import { BigNumber } from "ethers";
+import { Decryption } from "../typechain-types";
 import { deployDecryption } from "./tools/deploy/decryption";
 import { deployContractManager } from "./tools/deploy/contractManager";
-import { solidity } from "ethereum-waffle";
+import { fastBeforeEach } from "./tools/mocha";
 
 chai.should();
 chai.use(chaiAsPromised);
-chai.use(solidity);
 
 describe("Decryption", () => {
     let decryption: Decryption;
 
-    beforeEach(async () => {
+    fastBeforeEach(async () => {
         decryption = await deployDecryption(await deployContractManager());
     });
 
-    describe("when decryption contract is activated", async () => {
+    describe("when decryption contract is activated", () => {
         it("should encrypt and decrypt messages with a given key correctly", async () => {
             const secretNumber = 123456789;
             const key = "0x814eda04f881a67553ab65e4a0aeca015591a9aaa3f6bd2246508ce2f42905a6";
