@@ -1,23 +1,22 @@
 import { BigNumber } from "ethers";
 import { ContractManager,
          SkaleToken,
-         SkaleTokenInternalTester} from "../typechain";
+         SkaleTokenInternalTester} from "../typechain-types";
 
 import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
 import { deployContractManager } from "./tools/deploy/contractManager";
 import { deployValidatorService } from "./tools/deploy/delegation/validatorService";
 import { deploySkaleToken } from "./tools/deploy/skaleToken";
-import { deployReentrancyTester } from "./tools/deploy/test/reentracyTester";
+import { deployReentrancyTester } from "./tools/deploy/test/reentrancyTester";
 import { deploySkaleManagerMock } from "./tools/deploy/test/skaleManagerMock";
-import { ethers, web3 } from "hardhat";
+import { ethers } from "hardhat";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
 import { expect } from "chai";
-import { solidity } from "ethereum-waffle";
+import { fastBeforeEach } from "./tools/mocha";
 
 chai.should();
 chai.use(chaiAsPromised);
-chai.use(solidity);
 
 describe("SkaleToken", () => {
   let owner: SignerWithAddress;
@@ -29,10 +28,10 @@ describe("SkaleToken", () => {
   let skaleToken: SkaleToken;
   let contractManager: ContractManager;
 
-  const TOKEN_CAP: number = 7000000000;
+  const TOKEN_CAP = 7000000000;
   const TOTAL_SUPPLY = 5000000000;
 
-  beforeEach(async () => {
+  fastBeforeEach(async () => {
     [owner, holder, receiver, nilAddress, accountWith99] = await ethers.getSigners();
 
     contractManager = await deployContractManager();
@@ -138,7 +137,7 @@ describe("SkaleToken", () => {
   });
 
   it("holder balance should be bigger than 0 eth", async () => {
-    const holderBalance = await web3.eth.getBalance(holder.address);
+    const holderBalance = await ethers.provider.getBalance(holder.address);
     holderBalance.should.not.be.equal(0);
   });
 
@@ -210,7 +209,7 @@ describe("SkaleToken", () => {
     await skaleToken.mint(holder.address, amount, "0x", "0x");
 
     const reentrancyTester = await deployReentrancyTester(contractManager);
-    await reentrancyTester.prepareToReentracyCheck();
+    await reentrancyTester.prepareToReentrancyCheck();
 
     await skaleToken.connect(holder).transfer(reentrancyTester.address, amount)
       .should.be.eventually.rejectedWith("ReentrancyGuard: reentrant call");
@@ -240,7 +239,7 @@ describe("SkaleToken", () => {
   it("should parse call data correctly", async () => {
     const skaleTokenInternalTesterFactory = await ethers.getContractFactory("SkaleTokenInternalTester");
     const skaleTokenInternalTester = await skaleTokenInternalTesterFactory.deploy(contractManager.address, []) as SkaleTokenInternalTester;
-    await skaleTokenInternalTester.getMsgData().should.be.eventually.equal(web3.eth.abi.encodeFunctionSignature("getMsgData()"));
+    await skaleTokenInternalTester.getMsgData().should.be.eventually.equal(skaleTokenInternalTester.interface.encodeFunctionData("getMsgData"));
   });
 });
 
