@@ -1,11 +1,9 @@
 import { ContractManager,
     DelegationController,
     SkaleToken,
-    ValidatorService,
-    ConstantsHolder,
-    SkaleManagerMock} from "../../typechain";
+    ValidatorService} from "../../typechain-types";
 
-import { currentTime, skipTime } from "../tools/time";
+import { currentTime, nextMonth, skipTime } from "../tools/time";
 
 import * as chai from "chai";
 import chaiAsPromised from "chai-as-promised";
@@ -16,17 +14,14 @@ import { deploySkaleToken } from "../tools/deploy/skaleToken";
 import { deployTimeHelpersWithDebug } from "../tools/deploy/test/timeHelpersWithDebug";
 import { State } from "../tools/types";
 import { deployTimeHelpers } from "../tools/deploy/delegation/timeHelpers";
-import { deployConstantsHolder } from "../tools/deploy/constantsHolder";
-import { ethers, web3 } from "hardhat";
+import { ethers } from "hardhat";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/dist/src/signer-with-address";
 import { deploySkaleManagerMock } from "../tools/deploy/test/skaleManagerMock";
-import { solidity } from "ethereum-waffle";
 import { expect, assert } from "chai";
 import { makeSnapshot, applySnapshot } from "../tools/snapshot";
 
 chai.should();
 chai.use(chaiAsPromised);
-chai.use(solidity);
 
 
 describe("DelegationController", () => {
@@ -40,7 +35,6 @@ describe("DelegationController", () => {
     let skaleToken: SkaleToken;
     let delegationController: DelegationController;
     let validatorService: ValidatorService;
-    let constantsHolder: ConstantsHolder;
 
     const month = 60 * 60 * 24 * 31;
 
@@ -53,7 +47,6 @@ describe("DelegationController", () => {
         skaleToken = await deploySkaleToken(contractManager);
         delegationController = await deployDelegationController(contractManager);
         validatorService = await deployValidatorService(contractManager);
-        constantsHolder = await deployConstantsHolder(contractManager);
 
         const skaleManagerMock = await deploySkaleManagerMock(contractManager);
         await contractManager.setContractsAddress("SkaleManager", skaleManagerMock.address);
@@ -69,7 +62,7 @@ describe("DelegationController", () => {
         await applySnapshot(snapshot);
     });
 
-    describe("when arguments for delegation initialized", async () => {
+    describe("when arguments for delegation initialized", () => {
         let validatorId: number;
         let amount: number;
         let delegationPeriod: number;
@@ -123,10 +116,6 @@ describe("DelegationController", () => {
         it("should send request for delegation", async () => {
             amount = 100;
             await skaleToken.mint(holder1.address, amount, "0x", "0x");
-            // const { logs } = await delegationController.connect(holder1).delegate(
-            //     validatorId, amount, delegationPeriod, info);
-            // assert.equal(logs.length, 1, "No DelegationProposed Event emitted");
-            // assert.equal(logs[0].event, "DelegationProposed");
 
             await expect(
                 delegationController.connect(holder1).delegate(validatorId, amount, delegationPeriod, info)
@@ -163,7 +152,7 @@ describe("DelegationController", () => {
             await delegationController.connect(holder1).delegate(validatorId, amount, delegationPeriod, info);
         });
 
-        describe("when delegation request was created", async () => {
+        describe("when delegation request was created", () => {
             let validatorEnabled: number;
             before(async () => {
                 validatorEnabled = await makeSnapshot();
@@ -219,13 +208,13 @@ describe("DelegationController", () => {
             });
 
             it("should reject accepting request if next month started", async () => {
-                await skipTime(ethers, month);
+                await nextMonth(contractManager);
                 await delegationController.connect(validator).acceptPendingDelegation(delegationId)
                     .should.be.rejectedWith("The delegation request is outdated");
             });
 
             it("should reject accepting request if validator tried to accept request not assigned to him", async () => {
-                validatorService.connect(validator2).registerValidator(
+                await validatorService.connect(validator2).registerValidator(
                     "ValidatorName",
                     "Really good validator",
                     500,
@@ -258,14 +247,14 @@ describe("DelegationController", () => {
                 (await delegationController.getState(delegationId)).should.be.equal(State.PROPOSED);
             });
 
-            describe("when delegation is accepted", async () => {
+            describe("when delegation is accepted", () => {
                 let holder1DelegatedToValidator: number;
                 before(async () => {
                     delegationId = 0;
                     holder1DelegatedToValidator = await makeSnapshot();
                     await delegationController.connect(validator).acceptPendingDelegation(delegationId);
 
-                    await skipTime(ethers, month);
+                    await nextMonth(contractManager);
                 });
 
                 after(async () => {
@@ -275,7 +264,7 @@ describe("DelegationController", () => {
                 it("should allow validator to request undelegation", async () => {
                     await delegationController.connect(validator).requestUndelegation(delegationId);
 
-                    await skipTime(ethers, delegationPeriod * month);
+                    await nextMonth(contractManager, delegationPeriod);
 
                     (await delegationController.getState(delegationId)).should.be.equal(State.COMPLETED);
                     (await skaleToken.callStatic.getAndUpdateDelegatedAmount(holder1.address)).toNumber().should.be.equal(0);
@@ -293,7 +282,7 @@ describe("DelegationController", () => {
                     await delegationController.connect(validator2).requestUndelegation(delegationId)
                         .should.be.eventually.rejectedWith("Permission denied to request undelegation");
 
-                    await skipTime(ethers, delegationPeriod * month);
+                    await nextMonth(contractManager, delegationPeriod);
 
                     (await delegationController.getState(delegationId)).should.be.equal(State.DELEGATED);
                     (await skaleToken.callStatic.getAndUpdateDelegatedAmount(holder1.address)).toNumber().should.be.equal(amount);
@@ -306,18 +295,18 @@ describe("DelegationController", () => {
                     const twoDays = 2 * 24 * 60 * 60;
 
                     // skip time 2 days before delegation end
-                    await skipTime(ethers, delegationEndTimestamp - twoDays - await currentTime(web3));
+                    await skipTime(delegationEndTimestamp - twoDays - await currentTime());
 
                     await delegationController.connect(validator).requestUndelegation(delegationId)
                         .should.be.eventually.rejectedWith("Undelegation requests must be sent 3 days before the end of delegation period");
                     (await delegationController.getState(delegationId)).should.be.equal(State.DELEGATED);
 
-                    await skipTime(ethers, twoDays * 2);
+                    await skipTime(twoDays * 2);
 
                     await delegationController.connect(validator).requestUndelegation(delegationId);
                     (await delegationController.getState(delegationId)).should.be.equal(State.UNDELEGATION_REQUESTED);
 
-                    await skipTime(ethers, delegationPeriod * month);
+                    await nextMonth(contractManager, delegationPeriod);
 
                     (await delegationController.getState(delegationId)).should.be.equal(State.COMPLETED);
                     (await skaleToken.callStatic.getAndUpdateDelegatedAmount(holder1.address)).toNumber().should.be.equal(0);
