@@ -1,5 +1,5 @@
-const rewardPeriod: number = 3600;
-const deltaPeriod: number = 300;
-const checkTime: number = 120;
+const rewardPeriod = 3600;
+const deltaPeriod = 300;
+const checkTime = 120;
 
 export { rewardPeriod, deltaPeriod, checkTime };
