@@ -1,7 +1,6 @@
-import { ConstantsHolder, ContractManager } from "../../../typechain";
+import { ConstantsHolder, ContractManager } from "../../../typechain-types";
 import { deployFunctionFactory } from "./factory";
 
 const name = "ConstantsHolder";
 
-export const deployConstantsHolder: (contractManager: ContractManager) => Promise<ConstantsHolder>
-    = deployFunctionFactory(name);
+export const deployConstantsHolder = deployFunctionFactory(name) as (contractManager: ContractManager) => Promise<ConstantsHolder>;
