@@ -1,4 +1,4 @@
-import { ContractManager, DelegationController } from "../../../../typechain";
+import { ContractManager, DelegationController } from "../../../../typechain-types";
 import { deployBounty } from "../bounty";
 import { deployFunctionFactory } from "../factory";
 import { deploySkaleToken } from "../skaleToken";
@@ -6,14 +6,13 @@ import { deployDelegationPeriodManager } from "./delegationPeriodManager";
 import { deployTimeHelpers } from "./timeHelpers";
 import { deployValidatorService } from "./validatorService";
 
-const deployDelegationController: (contractManager: ContractManager) => Promise<DelegationController>
-    = deployFunctionFactory("DelegationController",
-                            async (contractManager: ContractManager) => {
-                                await deployValidatorService(contractManager);
-                                await deployTimeHelpers(contractManager);
-                                await deployDelegationPeriodManager(contractManager);
-                                await deploySkaleToken(contractManager);
-                                await deployBounty(contractManager);
-                            });
-
-export { deployDelegationController };
+export const deployDelegationController = deployFunctionFactory(
+    "DelegationController",
+    async (contractManager: ContractManager) => {
+        await deployValidatorService(contractManager);
+        await deployTimeHelpers(contractManager);
+        await deployDelegationPeriodManager(contractManager);
+        await deploySkaleToken(contractManager);
+        await deployBounty(contractManager);
+    }
+) as (contractManager: ContractManager) => Promise<DelegationController>;
