@@ -1,7 +1,6 @@
-import { ContractManager, DelegationPeriodManager } from "../../../../typechain";
+import { ContractManager, DelegationPeriodManager } from "../../../../typechain-types";
 import { deployFunctionFactory } from "../factory";
 
 const name = "DelegationPeriodManager";
 
-export const deployDelegationPeriodManager: (contractManager: ContractManager) => Promise<DelegationPeriodManager>
-    = deployFunctionFactory(name);
+export const deployDelegationPeriodManager = deployFunctionFactory(name) as (contractManager: ContractManager) => Promise<DelegationPeriodManager>;
