@@ -1,13 +1,12 @@
-import { ContractManager, Punisher } from "../../../../typechain";
+import { ContractManager, Punisher } from "../../../../typechain-types";
 import { deployFunctionFactory } from "../factory";
 import { deployDelegationController } from "./delegationController";
 import { deployValidatorService } from "./validatorService";
 
-const deployPunisher: (contractManager: ContractManager) => Promise<Punisher>
-    = deployFunctionFactory("Punisher",
-                            async (contractManager: ContractManager) => {
-                                await deployDelegationController(contractManager);
-                                await deployValidatorService(contractManager);
-                            });
-
-export { deployPunisher };
+export const deployPunisher = deployFunctionFactory(
+    "Punisher",
+    async (contractManager: ContractManager) => {
+        await deployDelegationController(contractManager);
+        await deployValidatorService(contractManager);
+    }
+) as (contractManager: ContractManager) => Promise<Punisher>;
