@@ -1,15 +1,13 @@
 import { ContractManager,
-    TimeHelpers } from "../../../../typechain";
+    TimeHelpers } from "../../../../typechain-types";
 import { deployWithConstructor, deployWithConstructorFunctionFactory } from "../factory";
 
 const name = "TimeHelpers";
 
-export const deployTimeHelpers: (contractManager: ContractManager) => Promise<TimeHelpers>
-    = deployWithConstructorFunctionFactory(
-        name,
-        async (_: ContractManager) => {
-            return undefined;
-        },
-        async (_: ContractManager) => {
-            return await deployWithConstructor(name);
-        });
+export const deployTimeHelpers = deployWithConstructorFunctionFactory(
+    name,
+    () => Promise.resolve(undefined),
+    async () => {
+        return await deployWithConstructor(name);
+    }
+) as (contractManager: ContractManager) => Promise<TimeHelpers>;
