@@ -1,14 +1,10 @@
-import { artifacts } from "hardhat";
-import { ContractManager, ECDH } from "../../../typechain";
+import { ContractManager, ECDH } from "../../../typechain-types";
 import { deployFunctionFactory, deployWithConstructor } from "./factory";
 
-const deployECDH: (contractManager: ContractManager) => Promise<ECDH>
-    = deployFunctionFactory("ECDH",
-                            async (contractManager: ContractManager) => {
-                                return undefined;
-                            },
-                            async (contractManager: ContractManager) => {
-                                return await deployWithConstructor("ECDH");
-                            });
-
-export { deployECDH };
+export const deployECDH = deployFunctionFactory(
+    "ECDH",
+    () => Promise.resolve(undefined),
+    async () => {
+        return await deployWithConstructor("ECDH");
+    }
+) as (contractManager: ContractManager) => Promise<ECDH>;
