@@ -1,9 +1,10 @@
 import { ethers } from "hardhat";
-import { ContractManager, SkaleToken } from "../../../typechain";
+import { ContractManager, SkaleToken } from "../../../typechain-types";
 import { deployDelegationController } from "./delegation/delegationController";
 import { deployPunisher } from "./delegation/punisher";
 import { deployTokenState } from "./delegation/tokenState";
 import { deployFunctionFactory } from "./factory";
+import { deploySkaleManager } from "./skaleManager";
 
 const name = "SkaleToken";
 
@@ -16,7 +17,11 @@ async function deployDependencies(contractManager: ContractManager) {
     await deployTokenState(contractManager);
     await deployDelegationController(contractManager);
     await deployPunisher(contractManager);
+    await deploySkaleManager(contractManager);
 }
 
-export const deploySkaleToken: (contractManager: ContractManager) => Promise<SkaleToken>
-    = deployFunctionFactory(name, deployDependencies, deploy);
+export const deploySkaleToken = deployFunctionFactory(
+    name,
+    deployDependencies,
+    deploy
+) as (contractManager: ContractManager) => Promise<SkaleToken>;
