@@ -1,6 +1,6 @@
-import { ContractManager, SlashingTable } from "../../../typechain";
+import { ContractManager, SlashingTable } from "../../../typechain-types";
 import { deployFunctionFactory } from "./factory";
 
 const name = "SlashingTable";
 
-export const deploySlashingTable: (contractManager: ContractManager) => Promise<SlashingTable> = deployFunctionFactory(name);
+export const deploySlashingTable = deployFunctionFactory(name) as (contractManager: ContractManager) => Promise<SlashingTable>;
