@@ -1,13 +1,10 @@
-import { ContractManager, PartialDifferencesTester } from "../../../../typechain";
+import { ContractManager, PartialDifferencesTester } from "../../../../typechain-types";
 import { deployWithConstructor, deployWithConstructorFunctionFactory } from "../factory";
 
-const deployPartialDifferencesTester: (contractManager: ContractManager) => Promise<PartialDifferencesTester>
-    = deployWithConstructorFunctionFactory("PartialDifferencesTester",
-                            async (_: ContractManager) => {
-                                return undefined;
-                            },
-                            async (_: ContractManager) => {
-                                return await deployWithConstructor("PartialDifferencesTester");
-                            });
-
-export { deployPartialDifferencesTester };
+export const deployPartialDifferencesTester = deployWithConstructorFunctionFactory(
+    "PartialDifferencesTester",
+    undefined,
+    async () => {
+        return await deployWithConstructor("PartialDifferencesTester");
+    }
+) as (contractManager: ContractManager) => Promise<PartialDifferencesTester>;
