@@ -1,7 +1,5 @@
-import { ContractManager, SkaleManagerMock } from "../../../../typechain";
+import { ContractManager, SkaleManagerMock } from "../../../../typechain-types";
 import { deployWithConstructorFunctionFactory } from "../factory";
 
-const deploySkaleManagerMock: (contractManager: ContractManager) => Promise<SkaleManagerMock>
-    = deployWithConstructorFunctionFactory("SkaleManagerMock");
-
-export { deploySkaleManagerMock };
+export const deploySkaleManagerMock
+    = deployWithConstructorFunctionFactory("SkaleManagerMock") as (contractManager: ContractManager) => Promise<SkaleManagerMock>;
