@@ -1,16 +1,14 @@
-import { ContractManager, Wallets } from "../../../typechain";
+import { ContractManager, Wallets } from "../../../typechain-types";
 import { deployValidatorService } from "./delegation/validatorService";
 import { deployFunctionFactory } from "./factory";
 import { deployNodes } from "./nodes";
 import { deploySchainsInternal } from "./schainsInternal";
 
-const deployWallets:
-    (contractManager: ContractManager) => Promise<Wallets>
-    = deployFunctionFactory("Wallets",
-                            async (contractManager: ContractManager) => {
-                                await deployNodes(contractManager);
-                                await deployValidatorService(contractManager);
-                                await deploySchainsInternal(contractManager);
-                            });
-
-export { deployWallets };
+export const deployWallets = deployFunctionFactory(
+    "Wallets",
+    async (contractManager: ContractManager) => {
+        await deployNodes(contractManager);
+        await deployValidatorService(contractManager);
+        await deploySchainsInternal(contractManager);
+    }
+) as (contractManager: ContractManager) => Promise<Wallets>;
