@@ -1,7 +1,7 @@
 import chai from "chai";
 import chaiAsPromised from "chai-as-promised";
-import { web3, ethers } from "hardhat"
-import { MathUtilsTester } from "../../typechain/MathUtilsTester";
+import { ethers } from "hardhat"
+import { MathUtilsTester } from "../../typechain-types/MathUtilsTester";
 import { makeSnapshot, applySnapshot } from "../tools/snapshot";
 
 chai.should();
@@ -23,7 +23,7 @@ describe("MathUtils", () => {
         await applySnapshot(snapshot);
     });
 
-    describe("in transaction", async () => {
+    describe("in transaction", () => {
         it("should subtract normally if reduced is greater than subtracted", async () => {
             (await mathUtils.callStatic.boundedSub(5, 3)).toNumber().should.be.equal(2);
             (await mathUtils.boundedSubWithoutEvent(5, 3)).toNumber().should.be.equal(2);
@@ -31,17 +31,16 @@ describe("MathUtils", () => {
 
         it("should return 0 if reduced is less than subtracted and emit event", async () => {
             (await mathUtils.callStatic.boundedSub(3, 5)).toNumber().should.be.equal(0);
-            const receipt = await (await mathUtils.boundedSub(3, 5)).wait();
-            receipt.logs.length.should.be.equal(1);
-            const log = receipt.logs[0];
-            log.topics.should.contain(web3.utils.keccak256("UnderflowError(uint256,uint256)"));
-            const params = web3.eth.abi.decodeParameters(["uint256", "uint256"], log.data);
-            params[0].should.be.equal("3");
-            params[1].should.be.equal("5");
+
+            const factory = await ethers.getContractFactory("MathUtils");
+            const mathUtilsLib = factory.attach(mathUtils.address);
+            await mathUtils.boundedSub(3, 5)
+                .should.emit(mathUtilsLib, "UnderflowError")
+                .withArgs(3, 5);
         });
     });
 
-    describe("in call", async () => {
+    describe("in call", () => {
         it("should subtract normally if reduced is greater than subtracted", async () => {
             (await mathUtils.boundedSubWithoutEvent(5, 3)).toNumber().should.be.equal(2);
         });
